add auto sql

+++ src/main/java/com/aib/websystem/FruitRepository.java
@@ -2,6 +2,6 @@ package com.aib.websystem;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface FruitRepository extends CrudRepository<Fruit, Integer> {
+public interface FruitRepository extends CrudRepository<Fruit, Long> {
 
 }
